split utils into separate files
