+++ test/Claim/EmissionsERC20.sol.ts
@@ -1,4 +1,4 @@
-import * as Util from "../Util";
+import * as Util from "../../utils";
 import chai from "chai";
 import { ethers } from "hardhat";
 import * as claimUtil from "./ClaimUtil";
@@ -9,7 +9,7 @@ import {
   paddedUInt32,
   paddedUInt256,
   sixZeros,
-} from "../Util";
+} from "../../utils";
 import type { ReadWriteTier } from "../../typechain/ReadWriteTier";
 import { BigNumber, Contract } from "ethers";
