+++ test/Sale/Sale.sol.ts
@@ -1,4 +1,4 @@
-import * as Util from "../Util";
+import * as Util from "../../utils";
 import chai from "chai";
 import { artifacts, ethers } from "hardhat";
 import type { Contract, ContractFactory } from "ethers";
@@ -16,7 +16,7 @@ import type {
   StartEvent,
   TimeoutEvent,
 } from "../../typechain/Sale";
-import { op } from "../Util";
+import { op } from "../../utils";
 import { ReserveToken } from "../../typechain/ReserveToken";
 import { ReadWriteTier } from "../../typechain/ReadWriteTier";
 import { RedeemableERC20Factory } from "../../typechain/RedeemableERC20Factory";
