+++ test/Tier/ReadWriteTier.sol.ts
@@ -10,7 +10,7 @@ import {
   blockNumbersToReport,
   assertError,
   getEventArgs,
-} from "../Util";
+} from "../../utils";
 import type { Contract } from "ethers";
 
 const { assert } = chai;
