+++ test/Tier/VerifyTier.sol.ts
@@ -1,4 +1,4 @@
-import * as Util from "../Util";
+import * as Util from "../../utils";
 import chai from "chai";
 import { ethers } from "hardhat";
 import type { VerifyTier } from "../../typechain/VerifyTier";
