--- test/Util.ts
@@ -1,796 +1,0 @@
-import { ethers, artifacts } from "hardhat";
-import chai from "chai";
-import type {
-  RedeemableERC20,
-  RedeemableERC20ConfigStruct,
-} from "../typechain/RedeemableERC20";
-import type {
-  ImplementationEvent as ImplementationEventRedeemableERC20Factory,
-  RedeemableERC20Factory,
-} from "../typechain/RedeemableERC20Factory";
-import type { CombineTier } from "../typechain/CombineTier";
-import type {
-  CombineTierFactory,
-  ImplementationEvent as ImplementationEventCombineTierFactory,
-} from "../typechain/CombineTierFactory";
-import type { Verify } from "../typechain/Verify";
-import type {
-  ImplementationEvent as ImplementationEventVerifyFactory,
-  VerifyFactory,
-} from "../typechain/VerifyFactory";
-import type { VerifyTier } from "../typechain/VerifyTier";
-import type {
-  ImplementationEvent as ImplementationEventVerifyTierFactory,
-  VerifyTierFactory,
-} from "../typechain/VerifyTierFactory";
-import { BigNumber } from "ethers";
-import type {
-  Contract,
-  BytesLike,
-  BigNumberish,
-  ContractTransaction,
-} from "ethers";
-import { concat, Hexable, hexlify, Result, zeroPad } from "ethers/lib/utils";
-import type { SignerWithAddress } from "@nomiclabs/hardhat-ethers/signers";
-import type { BoundsStruct } from "../typechain/VMStateBuilder";
-
-const { assert } = chai;
-
-export const CREATOR_FUNDS_RELEASE_TIMEOUT_TESTING = 100;
-export const MAX_RAISE_DURATION_TESTING = 100;
-
-const smartPoolManagerAddress = process.env.BALANCER_SMART_POOL_MANAGER;
-if (smartPoolManagerAddress) {
-  console.log(`using existing SmartPoolManager: ${smartPoolManagerAddress}`);
-}
-const balancerSafeMathAddress = process.env.BALANCER_SAFE_MATH;
-if (balancerSafeMathAddress) {
-  console.log(`using existing BalancerSafeMath: ${balancerSafeMathAddress}`);
-}
-const rightsManagerAddress = process.env.BALANCER_RIGHTS_MANAGER;
-if (rightsManagerAddress) {
-  console.log(`using existing RightsManager: ${rightsManagerAddress}`);
-}
-const bFactoryAddress = process.env.BALANCER_BFACTORY;
-if (bFactoryAddress) {
-  console.log(`using existing BFactory: ${bFactoryAddress}`);
-}
-const crpFactoryAddress = process.env.BALANCER_CRP_FACTORY;
-if (crpFactoryAddress) {
-  console.log(`using existing CRPFactory: ${crpFactoryAddress}`);
-}
-
-export const basicDeploy = async (name, libs) => {
-  const factory = await ethers.getContractFactory(name, {
-    libraries: libs,
-  });
-
-  const contract = await factory.deploy();
-
-  await contract.deployed();
-
-  return contract;
-};
-
-export interface Factories {
-  redeemableERC20Factory: RedeemableERC20Factory & Contract;
-}
-
-export const factoriesDeploy = async (): Promise<Factories> => {
-  const redeemableERC20FactoryFactory = await ethers.getContractFactory(
-    "RedeemableERC20Factory",
-    {}
-  );
-  const redeemableERC20Factory =
-    (await redeemableERC20FactoryFactory.deploy()) as RedeemableERC20Factory &
-      Contract;
-  await redeemableERC20Factory.deployed();
-
-  const { implementation: implementation0 } = (await getEventArgs(
-    redeemableERC20Factory.deployTransaction,
-    "Implementation",
-    redeemableERC20Factory
-  )) as ImplementationEventRedeemableERC20Factory["args"];
-  assert(
-    !(implementation0 === zeroAddress),
-    "implementation redeemableERC20 factory zero address"
-  );
-
-  return {
-    redeemableERC20Factory,
-  };
-};
-
-export const zeroAddress = ethers.constants.AddressZero;
-export const oneAddress = "0x0000000000000000000000000000000000000001";
-export const eighteenZeros = "000000000000000000";
-
-export const fourZeros = "0000";
-export const sixZeros = "000000";
-export const nineZeros = "000000000";
-export const tenZeros = "0000000000";
-export const sixteenZeros = "0000000000000000";
-
-export const ONE = ethers.BigNumber.from("1" + eighteenZeros);
-export const RESERVE_ONE = ethers.BigNumber.from("1" + sixZeros);
-
-export const RESERVE_MIN_BALANCE = ethers.BigNumber.from("1" + sixZeros);
-
-export const max_uint256 = ethers.BigNumber.from(
-  "0xFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF"
-);
-export const max_uint32 = ethers.BigNumber.from("0xffffffff");
-export const max_uint16 = ethers.BigNumber.from("0xffff");
-
-export const ALWAYS = 0;
-export const NEVER = max_uint256;
-
-export const fixedPointMul = (a: BigNumber, b: BigNumber): BigNumber =>
-  a.mul(b).div(ONE);
-export const fixedPointDiv = (a: BigNumber, b: BigNumber): BigNumber =>
-  a.mul(ONE).div(b);
-export const minBN = (a: BigNumber, b: BigNumber): BigNumber =>
-  a.lt(b) ? a : b;
-export const maxBN = (a: BigNumber, b: BigNumber): BigNumber =>
-  a.gt(b) ? a : b;
-
-export const determineReserveDust = (bPoolReserveBalance: BigNumber) => {
-  let dust = bPoolReserveBalance.mul(ONE).div(1e7).div(ONE);
-  if (dust.lt(RESERVE_MIN_BALANCE)) {
-    dust = RESERVE_MIN_BALANCE;
-  }
-  return dust;
-};
-
-export const assertError = async (f, s: string, e: string) => {
-  let didError = false;
-  try {
-    await f();
-  } catch (e) {
-    assert(e.toString().includes(s), `error string ${e} does not include ${s}`);
-    didError = true;
-  }
-  assert(didError, e);
-};
-
-export const verifyDeploy = async (deployer, config) => {
-  const factoryFactory = await ethers.getContractFactory("VerifyFactory");
-  const factory = (await factoryFactory.deploy()) as VerifyFactory;
-  await factory.deployed();
-
-  const { implementation } = (await getEventArgs(
-    factory.deployTransaction,
-    "Implementation",
-    factory
-  )) as ImplementationEventVerifyFactory["args"];
-  assert(
-    !(implementation === zeroAddress),
-    "implementation verify factory zero address"
-  );
-
-  const tx = await factory.createChildTyped(config);
-  const contract = new ethers.Contract(
-    ethers.utils.hexZeroPad(
-      ethers.utils.hexStripZeros(
-        (await getEventArgs(tx, "NewChild", factory)).child
-      ),
-      20
-    ),
-    (await artifacts.readArtifact("Verify")).abi,
-    deployer
-  ) as Verify & Contract;
-  await contract.deployed();
-  return contract;
-};
-
-export const verifyTierDeploy = async (deployer, config) => {
-  const factoryFactory = await ethers.getContractFactory("VerifyTierFactory");
-  const factory = (await factoryFactory.deploy()) as VerifyTierFactory;
-  await factory.deployed();
-  const tx = await factory.createChildTyped(config);
-  const contract = new ethers.Contract(
-    ethers.utils.hexZeroPad(
-      ethers.utils.hexStripZeros(
-        (await getEventArgs(tx, "NewChild", factory)).child
-      ),
-      20
-    ),
-    (await artifacts.readArtifact("VerifyTier")).abi,
-    deployer
-  ) as VerifyTier & Contract;
-  await contract.deployed();
-
-  const { implementation } = (await getEventArgs(
-    factory.deployTransaction,
-    "Implementation",
-    factory
-  )) as ImplementationEventVerifyTierFactory["args"];
-  assert(
-    !(implementation === zeroAddress),
-    "implementation verifyTier factory zero address"
-  );
-
-  return contract;
-};
-
-export const combineTierDeploy = async (deployer, config) => {
-  const stateBuilderFactory = await ethers.getContractFactory(
-    "AllStandardOpsStateBuilder"
-  );
-  const stateBuilder = await stateBuilderFactory.deploy();
-  await stateBuilder.deployed();
-
-  const factoryFactory = await ethers.getContractFactory("CombineTierFactory");
-  const factory = (await factoryFactory.deploy(
-    stateBuilder.address
-  )) as CombineTierFactory;
-  await factory.deployed();
-
-  const { implementation } = (await getEventArgs(
-    factory.deployTransaction,
-    "Implementation",
-    factory
-  )) as ImplementationEventCombineTierFactory["args"];
-  assert(
-    !(implementation === zeroAddress),
-    "implementation combineTier factory zero address"
-  );
-
-  const tx = await factory.createChildTyped(config);
-  const contract = new ethers.Contract(
-    ethers.utils.hexZeroPad(
-      ethers.utils.hexStripZeros(
-        (await getEventArgs(tx, "NewChild", factory)).child
-      ),
-      20
-    ),
-    (await artifacts.readArtifact("CombineTier")).abi,
-    deployer
-  ) as CombineTier & Contract;
-  await contract.deployed();
-
-  return contract;
-};
-
-export const redeemableERC20Deploy = async (
-  deployer: SignerWithAddress,
-  config: RedeemableERC20ConfigStruct
-) => {
-  const redeemableERC20FactoryFactory = await ethers.getContractFactory(
-    "RedeemableERC20Factory"
-  );
-  const redeemableERC20Factory =
-    (await redeemableERC20FactoryFactory.deploy()) as RedeemableERC20Factory;
-  await redeemableERC20Factory.deployed();
-
-  const { implementation } = (await getEventArgs(
-    redeemableERC20Factory.deployTransaction,
-    "Implementation",
-    redeemableERC20Factory
-  )) as ImplementationEventRedeemableERC20Factory["args"];
-  assert(
-    !(implementation === zeroAddress),
-    "implementation redeemableERC20 factory zero address"
-  );
-
-  const txDeploy = await redeemableERC20Factory.createChildTyped(config);
-  const redeemableERC20 = new ethers.Contract(
-    ethers.utils.hexZeroPad(
-      ethers.utils.hexStripZeros(
-        (await getEventArgs(txDeploy, "NewChild", redeemableERC20Factory)).child
-      ),
-      20
-    ),
-    (await artifacts.readArtifact("RedeemableERC20")).abi,
-    deployer
-  ) as RedeemableERC20 & Contract;
-
-  await redeemableERC20.deployed();
-
-  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-  // @ts-ignore
-  redeemableERC20.deployTransaction = txDeploy;
-
-  return redeemableERC20;
-};
-
-export const createEmptyBlock = async (count?: number): Promise<void> => {
-  const signers = await ethers.getSigners();
-  const tx = { to: signers[1].address };
-  if (count > 0) {
-    for (let i = 0; i < count; i++) {
-      await signers[0].sendTransaction(tx);
-    }
-  } else {
-    await signers[0].sendTransaction(tx);
-  }
-};
-
-/**
- * Utility function that transforms a hexadecimal number from the output of the ITier contract report
- * @param report String with Hexadecimal containing the array data
- * @returns number[] Block array of the reports
- */
-export function tierReport(report: string): number[] {
-  const parsedReport: number[] = [];
-  const arrStatus = [0, 1, 2, 3, 4, 5, 6, 7]
-    .map((i) =>
-      BigInt(report)
-        .toString(16)
-        .padStart(64, "0")
-        .slice(i * 8, i * 8 + 8)
-    )
-    .reverse();
-  //arrStatus = arrStatus.reverse();
-
-  for (const i in arrStatus) {
-    parsedReport.push(parseInt("0x" + arrStatus[i]));
-  }
-
-  return parsedReport;
-}
-
-export function blockNumbersToReport(blockNos: number[]): BigNumber {
-  assert(blockNos.length === 8);
-
-  return ethers.BigNumber.from(
-    "0x" +
-      [...blockNos]
-        .reverse()
-        .map((i) => BigInt(i).toString(16).padStart(8, "0"))
-        .join("")
-  );
-}
-
-/**
- * Pads leading zeroes of hex number to hex string length of 32 bytes
- * @param {BigNumber} hex
- */
-export function zeroPad32(hex: BigNumber): string {
-  return ethers.utils.hexZeroPad(hex.toHexString(), 32);
-}
-
-/**
- * Pads leading zeroes of hex number to hex string length of 4 bytes
- * @param {BigNumber} hex
- */
-export function zeroPad4(hex: BigNumber): string {
-  return ethers.utils.hexZeroPad(hex.toHexString(), 4);
-}
-
-/**
- * Converts a value to raw bytes representation. Assumes `value` is less than or equal to 1 byte, unless a desired `bytesLength` is specified.
- *
- * @param value - value to convert to raw bytes format
- * @param bytesLength - (defaults to 1) number of bytes to left pad if `value` doesn't completely fill the desired amount of memory. Will throw `InvalidArgument` error if value already exceeds bytes length.
- * @returns {Uint8Array} - raw bytes representation
- */
-export function bytify(
-  value: number | BytesLike | Hexable,
-  bytesLength = 1
-): BytesLike {
-  return zeroPad(hexlify(value), bytesLength);
-}
-
-/**
- * Constructs the operand for RainVM's `call` opcode by packing 3 numbers into a single byte. All parameters use zero-based counting i.e. an `fnSize` of 0 means to allocate one element (32 bytes) on the stack to define your functions, while an `fnSize` of 3 means to allocate all four elements (4 * 32 bytes) on the stack.
- *
- * @param sourceIndex - index of function source in `immutableSourceConfig.sources`
- * @param loopSize - number of times to subdivide vals, reduces uint size but allows for more vals (range 0-7)
- * @param valSize - number of vals in outer stack (range 0-7)
- */
-export function callSize(
-  sourceIndex: number,
-  loopSize: number,
-  valSize: number
-): number {
-  // CallSize(
-  //   op_.val & 0x07,      // 00000111
-  //   op_.val >> 3 & 0x03, // 00011000
-  //   op_.val >> 5 & 0x07  // 11100000
-  // )
-
-  if (sourceIndex < 0 || sourceIndex > 7) {
-    throw new Error("Invalid fnSize");
-  } else if (loopSize < 0 || loopSize > 3) {
-    throw new Error("Invalid loopSize");
-  } else if (valSize < 0 || valSize > 7) {
-    throw new Error("Invalid valSize");
-  }
-  let callSize = valSize;
-  callSize <<= 2;
-  callSize += loopSize;
-  callSize <<= 3;
-  callSize += sourceIndex;
-  return callSize;
-}
-
-export function skip(places: number, conditional = false): number {
-  let skip = conditional ? 1 : 0;
-  skip <<= 7;
-  // JS ints are already signed.
-  skip |= places & 0x7f;
-  return skip;
-}
-
-/**
- * Converts an opcode and operand to bytes, and returns their concatenation.
- * @param code - the opcode
- * @param erand - the operand, currently limited to 1 byte (defaults to 0)
- */
-export function op(
-  code: number,
-  erand: number | BytesLike | Hexable = 0
-): Uint8Array {
-  return concat([bytify(code), bytify(erand)]);
-}
-
-export const wrap2BitUInt = (integer: number) => {
-  while (integer > 3) {
-    integer -= 4;
-  }
-  return integer;
-};
-
-export const wrap4BitUInt = (integer: number) => {
-  while (integer > 15) {
-    integer -= 16;
-  }
-  return integer;
-};
-
-export const wrap8BitUInt = (integer: number) => {
-  while (integer > 255) {
-    integer -= 256;
-  }
-  return integer;
-};
-
-export const array2BitUInts = (length) =>
-  Array(length)
-    .fill(0)
-    // .map((_, i) => 3);
-    .map((_, i) => wrap2BitUInt(i));
-
-export const array4BitUInts = (length) =>
-  Array(length)
-    .fill(0)
-    .map((_, i) => wrap4BitUInt(i));
-
-export const array8BitUInts = (length) =>
-  Array(length)
-    .fill(0)
-    .map((_, i) => wrap8BitUInt(i));
-
-export const pack32UIntsIntoByte = (numArray: number[]): number[] => {
-  const val: number[] = [];
-  let valIndex = 0;
-
-  for (let i = 0; i < numArray.length; i += 4) {
-    const byte =
-      (numArray[i] << 6) +
-      (numArray[i + 1] << 4) +
-      (numArray[i + 2] << 2) +
-      numArray[i + 3];
-
-    val[valIndex] = byte;
-    valIndex++;
-  }
-
-  return val;
-};
-
-export const paddedUInt256 = (report: BigNumber): string => {
-  if (report.gt(max_uint256)) {
-    throw new Error(`${report} exceeds max uint256`);
-  }
-  return "0x" + report.toHexString().substring(2).padStart(64, "0");
-};
-
-export const paddedUInt32 = (number: number | BytesLike | Hexable): string => {
-  if (ethers.BigNumber.from(number).gt(max_uint32)) {
-    throw new Error(`${number} exceeds max uint32`);
-  }
-  return hexlify(number).substring(2).padStart(8, "0");
-};
-
-export type Source = [BigNumberish, BigNumberish, BigNumberish, BigNumberish];
-export type Constants = [
-  BigNumberish,
-  BigNumberish,
-  BigNumberish,
-  BigNumberish,
-  BigNumberish,
-  BigNumberish,
-  BigNumberish,
-  BigNumberish,
-  BigNumberish,
-  BigNumberish,
-  BigNumberish,
-  BigNumberish,
-  BigNumberish,
-  BigNumberish,
-  BigNumberish,
-  BigNumberish
-];
-
-/**
- *
- * @param tx - transaction where event occurs
- * @param eventName - name of event
- * @param contract - contract object holding the address, filters, interface
- * @param contractAddressOverride - (optional) override the contract address which emits this event
- * @returns Array of events with their arguments, which can each be deconstructed by array index or by object key
- */
-export const getEvents = async (
-  tx: ContractTransaction,
-  eventName: string,
-  contract: Contract,
-  contractAddressOverride: string = null
-): Promise<Result[]> => {
-  const address = contractAddressOverride
-    ? contractAddressOverride
-    : contract.address;
-
-  const eventObjs = (await tx.wait()).events.filter(
-    (x) =>
-      x.topics[0] == contract.filters[eventName]().topics[0] &&
-      x.address == address
-  );
-
-  if (!eventObjs.length) {
-    throw new Error(`Could not find event ${eventName} at address ${address}`);
-  }
-
-  return eventObjs.map((eventObj) =>
-    contract.interface.decodeEventLog(eventName, eventObj.data)
-  );
-};
-
-/**
- *
- * @param tx - transaction where event occurs
- * @param eventName - name of event
- * @param contract - contract object holding the address, filters, interface
- * @param contractAddressOverride - (optional) override the contract address which emits this event
- * @returns Event arguments of first matching event, can be deconstructed by array index or by object key
- */
-export const getEventArgs = async (
-  tx: ContractTransaction,
-  eventName: string,
-  contract: Contract,
-  contractAddressOverride: string = null
-): Promise<Result> => {
-  return (await getEvents(tx, eventName, contract, contractAddressOverride))[0];
-};
-
-export function selectLte(logic: number, mode: number, length: number): number {
-  let lte = logic;
-  lte <<= 2;
-  lte += mode;
-  lte <<= 5;
-  lte += length;
-  return lte;
-}
-
-export enum selectLteLogic {
-  every,
-  any,
-}
-
-export enum selectLteMode {
-  min,
-  max,
-  first,
-}
-
-export enum AllStandardOps {
-  CONSTANT,
-  STACK,
-  CONTEXT,
-  STORAGE,
-  ZIPMAP,
-  DEBUG,
-  BLOCK_NUMBER,
-  BLOCK_TIMESTAMP,
-  SENDER,
-  THIS_ADDRESS,
-  SCALE18_MUL,
-  SCALE18_DIV,
-  SCALE18,
-  SCALEN,
-  SCALE_BY,
-  ADD,
-  SATURATING_ADD,
-  SUB,
-  SATURATING_SUB,
-  MUL,
-  SATURATING_MUL,
-  DIV,
-  MOD,
-  EXP,
-  MIN,
-  MAX,
-  ISZERO,
-  EAGER_IF,
-  EQUAL_TO,
-  LESS_THAN,
-  GREATER_THAN,
-  EVERY,
-  ANY,
-  REPORT,
-  SATURATING_DIFF,
-  UPDATE_BLOCKS_FOR_TIER_RANGE,
-  SELECT_LTE,
-  IERC20_BALANCE_OF,
-  IERC20_TOTAL_SUPPLY,
-  IERC721_BALANCE_OF,
-  IERC721_OWNER_OF,
-  IERC1155_BALANCE_OF,
-  IERC1155_BALANCE_OF_BATCH,
-  length,
-}
-
-const testStructs = (
-  solObj: Record<string, unknown>,
-  jsObj: Record<string, unknown>
-) => {
-  Object.keys(solObj).forEach((key) => {
-    let expectedValue = jsObj[key];
-    let actualValue = solObj[key];
-
-    if (expectedValue !== undefined) {
-      if (expectedValue instanceof Uint8Array) {
-        expectedValue = hexlify(expectedValue);
-      }
-      if (actualValue instanceof BigNumber) {
-        expectedValue = BigNumber.from(expectedValue);
-      }
-
-      if (
-        typeof actualValue === "string" ||
-        typeof expectedValue === "string"
-      ) {
-        actualValue = `${actualValue}`.toLowerCase();
-        expectedValue = `${expectedValue}`.toLowerCase();
-      }
-
-      if (
-        typeof actualValue === "object" ||
-        typeof expectedValue === "object"
-      ) {
-        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-        // @ts-ignore
-        // recursive call for nested structs
-        testStructs(actualValue, expectedValue);
-      } else {
-        let condition: boolean;
-        try {
-          condition =
-            actualValue == expectedValue || actualValue["eq"](expectedValue);
-        } catch (error) {
-          console.log(error);
-        }
-
-        assert(
-          condition,
-          `wrong value for property: '${key}'
-          expected  ${expectedValue}
-          got       ${actualValue}`
-        );
-      }
-    }
-  });
-};
-
-/**
- * Uses chai `assert` to compare a Solidity struct with a JavaScript object by checking whether the values for each property are equivalent.
- * Will safely recurse over nested structs and compare nested properties.
- * Throws an error if any comparisons fail.
- * @param solStruct - Solidity struct, returned from something such as an emitted solidity Event. This should have an array-like structure with raw values followed by key-values (e.g. `solStruct: ['foo', 'bar', prop1: 'foo', prop2: 'bar']`).
- * @param jsObj - JavaScript object literal to use as comparison.
- */
-export const compareStructs = (
-  solStruct: unknown[],
-  jsObj: Record<string, unknown>
-) => {
-  const solEntries = Object.entries(solStruct).splice(
-    solStruct.length // actually half the solStruct size
-  );
-
-  if (!solEntries.length) {
-    throw new Error(
-      `Could not generate entries from a solStruct of length ${solStruct.length}. Ensure you are using a Solidity struct for solStruct.`
-    );
-  }
-
-  const solObj = Object.fromEntries(solEntries);
-
-  testStructs(solObj, jsObj);
-};
-
-const testSolStructs = (
-  solActualObj: Record<string, unknown>,
-  solExpectedObj: Record<string, unknown>
-) => {
-  Object.keys(solActualObj).forEach((key) => {
-    const actualValue = solActualObj[key];
-    const expectedValue = solExpectedObj[key];
-
-    if (typeof actualValue === "object" || typeof expectedValue === "object") {
-      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-      // @ts-ignore
-      // recursive call for nested structs
-      testSolStructs(actualValue, expectedValue);
-    } else {
-      let condition: boolean;
-      try {
-        condition =
-          actualValue == expectedValue || actualValue["eq"](expectedValue);
-      } catch (error) {
-        console.log(error);
-      }
-
-      assert(
-        condition,
-        `wrong value for property: '${key}'
-        expected  ${expectedValue}
-        got       ${actualValue}`
-      );
-    }
-  });
-};
-
-/**
- * Uses chai `assert` to compare Solidity structs by checking whether the values for each property are equivalent.
- * Will safely recurse over nested structs and compare nested properties.
- * Throws an error if any comparisons fail.
- * @param solStructActual - Solidity struct, returned from something such as an emitted solidity Event. This should have an array-like structure with raw values followed by key-values (e.g. `solStruct: ['foo', 'bar', prop1: 'foo', prop2: 'bar']`).
- * @param solStructExpected - Solidity struct.
- */
-export const compareSolStructs = (
-  solStructActual: unknown[],
-  solStructExpected: unknown[]
-) => {
-  const solActualEntries = Object.entries(solStructActual).splice(
-    solStructActual.length // actually half the solStruct size
-  );
-  const solExpectedEntries = Object.entries(solStructExpected).splice(
-    solStructExpected.length // actually half the solStruct size
-  );
-
-  if (!solActualEntries.length) {
-    throw new Error(
-      `Could not generate entries from a solStructActual of length ${solStructActual.length}. Ensure you are using a Solidity struct for solStructActual.`
-    );
-  }
-  if (!solExpectedEntries.length) {
-    throw new Error(
-      `Could not generate entries from a solStructExpected of length ${solStructExpected.length}. Ensure you are using a Solidity struct for solStructExpected.`
-    );
-  }
-
-  const solAObj = Object.fromEntries(solActualEntries);
-  const solBObj = Object.fromEntries(solExpectedEntries);
-
-  testSolStructs(solAObj, solBObj);
-};
-
-export enum Debug {
-  StateAbi,
-  StatePacked,
-  Stack,
-  StackIndex,
-}
-
-export const newVMStateBuilderBounds = (): BoundsStruct => {
-  return {
-    entrypoint: 0,
-    minFinalStackIndex: 0,
-    stackIndex: 0,
-    stackLength: 0,
-    argumentsLength: 0,
-    storageLength: 0,
-    opcodesLength: 0,
-  };
-};
