+++ test/VM/LogicOpsTest.sol.ts
@@ -1,8 +1,13 @@
 import chai from "chai";
 import { ethers } from "hardhat";
-import { op, AllStandardOps } from "../Util";
+import { op, AllStandardOps } from "../../utils";
 import { concat, hexlify } from "ethers/lib/utils";
-import { callSize, max_uint256, paddedUInt256, paddedUInt32 } from "../Util";
+import {
+  callSize,
+  max_uint256,
+  paddedUInt256,
+  paddedUInt32,
+} from "../../utils";
 import type { BigNumber, Contract } from "ethers";
 import type { AllStandardOpsTest } from "../../typechain/AllStandardOpsTest";
