+++ utils/types.ts
@@ -1,0 +1,86 @@
+import { BigNumberish } from "ethers";
+
+export enum selectLteLogic {
+  every,
+  any,
+}
+
+export enum selectLteMode {
+  min,
+  max,
+  first,
+}
+
+export enum AllStandardOps {
+  CONSTANT,
+  STACK,
+  CONTEXT,
+  STORAGE,
+  ZIPMAP,
+  DEBUG,
+  BLOCK_NUMBER,
+  BLOCK_TIMESTAMP,
+  SENDER,
+  THIS_ADDRESS,
+  SCALE18_MUL,
+  SCALE18_DIV,
+  SCALE18,
+  SCALEN,
+  SCALE_BY,
+  ADD,
+  SATURATING_ADD,
+  SUB,
+  SATURATING_SUB,
+  MUL,
+  SATURATING_MUL,
+  DIV,
+  MOD,
+  EXP,
+  MIN,
+  MAX,
+  ISZERO,
+  EAGER_IF,
+  EQUAL_TO,
+  LESS_THAN,
+  GREATER_THAN,
+  EVERY,
+  ANY,
+  REPORT,
+  SATURATING_DIFF,
+  UPDATE_BLOCKS_FOR_TIER_RANGE,
+  SELECT_LTE,
+  IERC20_BALANCE_OF,
+  IERC20_TOTAL_SUPPLY,
+  IERC721_BALANCE_OF,
+  IERC721_OWNER_OF,
+  IERC1155_BALANCE_OF,
+  IERC1155_BALANCE_OF_BATCH,
+  length,
+}
+
+export enum Debug {
+  StateAbi,
+  StatePacked,
+  Stack,
+  StackIndex,
+}
+
+export type Source = [BigNumberish, BigNumberish, BigNumberish, BigNumberish];
+export type Constants = [
+  BigNumberish,
+  BigNumberish,
+  BigNumberish,
+  BigNumberish,
+  BigNumberish,
+  BigNumberish,
+  BigNumberish,
+  BigNumberish,
+  BigNumberish,
+  BigNumberish,
+  BigNumberish,
+  BigNumberish,
+  BigNumberish,
+  BigNumberish,
+  BigNumberish,
+  BigNumberish
+];
